Move classes back
